split up database files
